+++ R/predict-fit.R
@@ -35,6 +35,13 @@ tidypredict_fit.pm_tree <- function(model) {
       }
       return(generate_nested_case_when_tree(model$tree_info))
     }
+    if (model_type == "blackboost") {
+      return(mboost_build_formula(
+        model$tree_info_list,
+        model$general$nu,
+        model$general$offset
+      ))
+    }
     if (model_type %in% c("ranger", "randomForest", "cforest", "aorsf")) {
       # For forests, average all trees
       tree_exprs <- map(model$tree_info_list, generate_nested_case_when_tree)